added meaningful logs; prevent auto overwriting input.json

+++ src/aosm/azext_aosm/custom.py
@@ -221,7 +221,14 @@ def _generate_config(definition_type: str, output_file: str = "input.json"):
     """
     config = get_configuration(definition_type)
     config_as_dict = json.dumps(asdict(config), indent=4)
-
+    
+    if os.path.exists(output_file):
+        carry_on = input(
+            f"The file {output_file} already exists - do you want to overwrite it? (y/n)"
+        )
+        if carry_on != "y":
+            raise UnclassifiedUserFault("User aborted!")
+        
     with open(output_file, "w", encoding="utf-8") as f:
         f.write(config_as_dict)
         if definition_type == CNF or definition_type == VNF:
